pyproject.toml
[project]
name = "creative-variants"
version = "0.1.0"
description = "CLI tool for generating appearance variants of UGC videos"
authors = [{ name = "Unreal Labs" }]
requires-python = ">=3.11,<3.13"
dependencies = [
    "python-dotenv>=1.0",
    "PyJWT>=2.8",
    "typer>=0.9",
    "pydantic>=2.3",
    "requests>=2.31",
    "opencv-python>=4.8",
    "Pillow>=10.0",
    "numpy>=1.24",
    "pytesseract>=0.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
]

[project.scripts]
creative_variants = "src.main:app"

[tool.uv]
dev-dependencies = [
    "pytest>=7.4",
]

[tool.uv.sources]

[build-system]
requires = ["hatchling>=1.18"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.hatch.build.targets.sdist]
include = ["src", "README.md", "pyproject.toml", "env.example", "tests"]
